MB-36804: Ensure flush_mr_partial() obeys node size quota
(Backport of MB-36424 to alice).

When couchstore writes out modified node elements during document
saves, it it supposed to limit the number of bytes written in a single
node to a _chunk_threshold - by default 1279 bytes. If the node is
larger than the limit it should be split into multiple sibling nodes.

However, this limit is _not_ respected, resulting in overly-large
nodes being written out. In the case of the by-seqno B-Tree (which
always writes values to the rightmost leaf as seqnos are increasing),
it results in all leaf elements residing in a single leaf
node. Moreover, this means that adding another element to the B-Tree
effectively re-writes the entire tree, resulting in massive Write
Amplification. Local testing saw Write Amplification up to 743x,
compared to around 5x immediately after compaction (when B-Tree
structure is repaired).

Problem is in the handling of the node quota in flush_mr_partial() -
the while loop terminating conditions is if mr_quota goes below zero
(i.e. quota has been consumed). mr_quota is provided as an input
parameter, and then decremented inside the while loop.

However, the type of mr_quota is an unsigned size_t - as such when it
is decremented in the while loop after writing the key/value pair it
goes to a number close to SIZE_T_MAX, not a negative value. As a
result the while loop is not exited.

This bug has been present since July 2012 - was introduced via
http://review.couchbase.org/#/c/18008/

Reviewed-on: http://review.couchbase.org/117555
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>

+++ src/btree_modify.cc
@@ -12,7 +12,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-static couchstore_error_t flush_mr_partial(couchfile_modify_result *res, size_t mr_quota);
+static couchstore_error_t flush_mr_partial(couchfile_modify_result *res, int mr_quota);
 static couchstore_error_t flush_mr(couchfile_modify_result *res);
 static couchstore_error_t purge_node(couchfile_modify_request *rq,
                                       node_pointer *nptr,
@@ -202,7 +202,7 @@ static couchstore_error_t flush_mr(couchfile_modify_result *res)
 
 //Write a node using enough items from the values list to create a node
 //with uncompressed size of at least mr_quota
-static couchstore_error_t flush_mr_partial(couchfile_modify_result *res, size_t mr_quota)
+static couchstore_error_t flush_mr_partial(couchfile_modify_result *res, int mr_quota)
 {
     char *dst;
     couchstore_error_t errcode = COUCHSTORE_SUCCESS;
